refactor: reuse cvss schemas between csaf 2.0 and 2.1
