+++ lib/shared/csafAjv.js
@@ -1,8 +1,8 @@
 import addFormats from 'ajv-formats'
 import Ajv2020 from 'ajv/dist/2020.js'
-import cvss_v2_0 from './csafAjv/cvss-v2.0.js'
-import cvss_v3_0 from './csafAjv/cvss-v3.0.js'
-import cvss_v3_1 from './csafAjv/cvss-v3.1.js'
+import cvss_v2_0 from '../../schemas/cvss-v2.0.js'
+import cvss_v3_0 from '../../schemas/cvss-v3.0.js'
+import cvss_v3_1 from '../../schemas/cvss-v3.1.js'
 
 const csafAjv = new Ajv2020({ strict: false, allErrors: true })
 addFormats(csafAjv)
